+++ Makefile
@@ -25,6 +25,7 @@ SRC_FILES = main.c\
 			signals.c\
 			ll_utils2.c\
 			execution_utils.c\
+			execution_utils2.c\
 			src/parse/tokenizer.c\
 			src/parse/ft_strtok.c\
 			src/parse/init_token.c\
