+++ src/parse/token_to_array.c
@@ -30,6 +30,6 @@ char **tokens_to_array(t_token *token_list) {
     }
     array[index] = NULL; // Terminate the array with NULL
 
-    return array;
+    return (array);
 }
